Move components to *.jsx
