+++ src/components/StatusBar/ColorSchemeSelector/ColorSchemeSelector.jsx
