+++ src/components/StatusBar/EncodingSelector/EncodingSelector.jsx
