+++ src/components/StatusBar/LineSeparatorSelector/LineSeparatorSelector.jsx
